# Improve performance of multipolygon validation