--- a/geom/type_polygon.go
+++ b/geom/type_polygon.go
@@ -252,12 +252,12 @@ func (p Polygon) Envelope() (Envelope, bool) {
 	return p.outer.Envelope()
 }
 
+func (p Polygon) rings() []LineString {
+	dst := make([]LineString, 0, 1+len(p.holes))
+	return p.appendRings(dst)
+}
+
 func (p Polygon) appendRings(dst []LineString) []LineString {
-	if len(dst) == 0 && cap(dst) == 0 {
-		// Optimise for the exact length if it looks like the caller isn't
-		// re-using a LineString buffer.
-		dst = make([]LineString, 0, 1+len(p.holes))
-	}
 	dst = append(dst, p.outer)
 	dst = append(dst, p.holes...)
 	return dst
@@ -285,7 +285,7 @@ func (p Polygon) AsBinary(w io.Writer) error {
 	marsh := newWKBMarshaller(w)
 	marsh.writeByteOrder()
 	marsh.writeGeomType(wkbGeomTypePolygon)
-	rings := p.appendRings(nil)
+	rings := p.rings()
 	marsh.writeCount(len(rings))
 	for _, ring := range rings {
 		numPts := ring.NumPoints()
@@ -312,7 +312,7 @@ func (p Polygon) MarshalJSON() ([]byte, error) {
 // Coordinates returns the coordinates of the rings making up the Polygon
 // (external ring first, then internal rings after).
 func (p Polygon) Coordinates() [][]Coordinates {
-	rings := p.appendRings(nil)
+	rings := p.rings()
 	coords := make([][]Coordinates, len(rings))
 	for i, r := range rings {
 		n := r.NumPoints()