Main changes:

- Use linesweep algorithm when determining which (sub)polygons to check for intersection. The old algorithm was the naive O(n^2) "check every pair" strategy.
- Some micro optimisations, mainly relating to reducing memory allocations.

Perf diff:

```
benchmark                                                 old ns/op      new ns/op     delta
BenchmarkIntersectsLineStringWithLineString/n=10-2        1832           1894          +3.38%
BenchmarkIntersectsLineStringWithLineString/n=100-2       16316          16613         +1.82%
BenchmarkIntersectsLineStringWithLineString/n=1000-2      194317         195022        +0.36%
BenchmarkIntersectsLineStringWithLineString/n=10000-2     2160753        2165766       +0.23%
BenchmarkPolygonSingleRingValidation/n=10-2               2171           1972          -9.17%
BenchmarkPolygonSingleRingValidation/n=100-2              24628          21594         -12.32%
BenchmarkPolygonSingleRingValidation/n=1000-2             285073         272031        -4.57%
BenchmarkPolygonSingleRingValidation/n=10000-2            3437431        3218969       -6.36%
BenchmarkPolygonMultipleRingsValidation/n=4-2             13837          13174         -4.79%
BenchmarkPolygonMultipleRingsValidation/n=36-2            229384         235612        +2.72%
BenchmarkPolygonMultipleRingsValidation/n=400-2           7882748        8637718       +9.58%
BenchmarkPolygonMultipleRingsValidation/n=4096-2          242356506      238706411     -1.51%
BenchmarkMultipolygonValidation/n=1-2                     1234           1429          +15.80%
BenchmarkMultipolygonValidation/n=4-2                     74022          15137         -79.55%
BenchmarkMultipolygonValidation/n=16-2                    1502411        136865        -90.89%
BenchmarkMultipolygonValidation/n=64-2                    25203805       1196663       -95.25%
BenchmarkMultipolygonValidation/n=256-2                   374222039      9820994       -97.38%
BenchmarkMultipolygonValidation/n=1024-2                  6088898206     78850656      -98.71%

benchmark                                                 old allocs     new allocs     delta
BenchmarkIntersectsLineStringWithLineString/n=10-2        21             21             +0.00%
BenchmarkIntersectsLineStringWithLineString/n=100-2       21             21             +0.00%
BenchmarkIntersectsLineStringWithLineString/n=1000-2      21             21             +0.00%
BenchmarkIntersectsLineStringWithLineString/n=10000-2     21             21             +0.00%
BenchmarkPolygonSingleRingValidation/n=10-2               18             14             -22.22%
BenchmarkPolygonSingleRingValidation/n=100-2              21             14             -33.33%
BenchmarkPolygonSingleRingValidation/n=1000-2             24             14             -41.67%
BenchmarkPolygonSingleRingValidation/n=10000-2            34             14             -58.82%
BenchmarkPolygonMultipleRingsValidation/n=4-2             169            159            -5.92%
BenchmarkPolygonMultipleRingsValidation/n=36-2            3096           3022           -2.39%
BenchmarkPolygonMultipleRingsValidation/n=400-2           114274         113472         -0.70%
BenchmarkPolygonMultipleRingsValidation/n=4096-2          3463225        3455030        -0.24%
BenchmarkMultipolygonValidation/n=1-2                     20             23             +15.00%
BenchmarkMultipolygonValidation/n=4-2                     1138           129            -88.66%
BenchmarkMultipolygonValidation/n=16-2                    21534          918            -95.74%
BenchmarkMultipolygonValidation/n=64-2                    357992         7089           -98.02%
BenchmarkMultipolygonValidation/n=256-2                   5781898        55956          -99.03%
BenchmarkMultipolygonValidation/n=1024-2                  92726799       444951         -99.52%

benchmark                                                 old bytes      new bytes     delta
BenchmarkIntersectsLineStringWithLineString/n=10-2        1216           1216          +0.00%
BenchmarkIntersectsLineStringWithLineString/n=100-2       7040           7040          +0.00%
BenchmarkIntersectsLineStringWithLineString/n=1000-2      66176          66176         +0.00%
BenchmarkIntersectsLineStringWithLineString/n=10000-2     656000         656000        +0.00%
BenchmarkPolygonSingleRingValidation/n=10-2               1352           680           -49.70%
BenchmarkPolygonSingleRingValidation/n=100-2              9336           4376          -53.13%
BenchmarkPolygonSingleRingValidation/n=1000-2             73976          41240         -44.25%
BenchmarkPolygonSingleRingValidation/n=10000-2            1843449        409880        -77.77%
BenchmarkPolygonMultipleRingsValidation/n=4-2             7072           6592          -6.79%
BenchmarkPolygonMultipleRingsValidation/n=36-2            126496         122944        -2.81%
BenchmarkPolygonMultipleRingsValidation/n=400-2           4552129        4513633       -0.85%
BenchmarkPolygonMultipleRingsValidation/n=4096-2          136951555      136558108     -0.29%
BenchmarkMultipolygonValidation/n=1-2                     728            728           +0.00%
BenchmarkMultipolygonValidation/n=4-2                     37008          4584          -87.61%
BenchmarkMultipolygonValidation/n=16-2                    691920         31976         -95.38%
BenchmarkMultipolygonValidation/n=64-2                    11467728       238248        -97.92%
BenchmarkMultipolygonValidation/n=256-2                   185069520      1837352       -99.01%
BenchmarkMultipolygonValidation/n=1024-2                  2967453936     14426672      -99.51%
```